Stack Point Implementation
This is an implementation of Stack with Pointers using Dynamic Memory.
It’s still in a working progress whether or not I want to use use
advanced arrays and pointers in it or not.  It all depends on the
purpose.  If anybody found this interesting, hit me up and we can talk
about expanding on this at a convenient time.